third_party/WebKit/LayoutTests/fast/viewport/viewport-68-expected.txt

CONSOLE WARNING: line 3: The value "100x" for key "width" was truncated to its numeric prefix.
ALERT: viewport size 320x352 scale 1 with limits [1, 5] and userScalable true
